+++ lib/src/fx.dart
@@ -186,6 +186,10 @@ class Fx<T> extends Iterable<T> {
 
   T? head() => s.head(_inner);
 
+  T? minBy(Object? Function(T a) f) => s.minBy(f, _inner);
+
+  T? maxBy(Object? Function(T a) f) => s.maxBy(f, _inner);
+
   (List<T>, List<T>) partition(bool Function(T a) f) => s.partition(f, _inner);
 
   int size() => s.size(_inner);
@@ -338,6 +342,10 @@ class FxAsync<T> implements FxAsyncIterable<T> {
   Future<T?> head() => s.headAsync(_inner);
 
   Future<T?> last() => s.lastAsync(_inner);
+
+  Future<T?> minBy(Object? Function(T a) f) => s.minByAsync(f, _inner);
+
+  Future<T?> maxBy(Object? Function(T a) f) => s.maxByAsync(f, _inner);
 
   Future<(List<T>, List<T>)> partition(FutureOr<bool> Function(T a) f) =>
       s.partitionAsync(f, _inner);